+++ server/migrations/20180813123650-create-customer.js
@@ -35,16 +35,8 @@ module.exports = {
         defaultValue: 0,
         allowNull: false,
       },
-      createdAt: {
-        allowNull: false,
-        defaultValue: Sequelize.fn('now'),
-        type: Sequelize.DATE,
-      },
-      updatedAt: {
-        allowNull: false,
-        defaultValue: Sequelize.fn('now'),
-        type: Sequelize.DATE,
-      },
+      createdAt: Sequelize.DATE,
+      updatedAt: Sequelize.DATE,
     })
   },
   down: (queryInterface, Sequelize) => {