+++ server/migrations/20180813123652-create-location.js
@@ -14,16 +14,8 @@ module.exports = {
       name: {
         type: Sequelize.STRING,
       },
-      createdAt: {
-        allowNull: false,
-        defaultValue: Sequelize.fn('now'),
-        type: Sequelize.DATE,
-      },
-      updatedAt: {
-        allowNull: false,
-        defaultValue: Sequelize.fn('now'),
-        type: Sequelize.DATE,
-      },
+      createdAt: Sequelize.DATE,
+      updatedAt: Sequelize.DATE,
     })
   },
   down: (queryInterface, Sequelize) => {