+++ src/store/customer/getters.js
@@ -3,7 +3,3 @@ export const getIsAuth = state => !!state.token
 export const getIsLoading = state => {
   return state.isLoading
 }
-
-export const getIsHiddenRegBtn = state => {
-  return state.isHiddenRegBtn
-}